refactored on bmesh_invert_normals.py and bmesh_recalculate_face_normals.pn
